# hashing: fix performance regression