benchmarks before:
```
 consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 36
 consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 735
 consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 32    datapoints: 100000 secs: 2.18479
 consistent-hashing  deplication_factor: 1 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 secs: 5.39408
```

benchmarks after:
```
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     datapoints: 100000 msecs: 44
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    datapoints: 100000 msecs: 483
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    datapoints: 100000 msecs: 871
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    datapoints: 100000 secs: 1.36102
  
```