@iksaif 2018-01-08
https://github.com/graphite-project/graphite-web/pull/2191