@deniszh 2018-01-08
It will not touch hashing order, right?