@iksaif 2018-01-08
> It will not touch hashing order, right?

sets are not ordered, so order in which the nodes would have been returned will change. I though about that, but I also did not find when this could be an issue (since all returned nodes are always used). Did I miss anything ?