@deniszh 2018-01-08
Well. We have not perfect test coverage, but current tests are green, so, let's assuming that everything is good. 👍