@DanCech 2018-01-08
Yeah, looking at how it's used it seems like it should be fine.